requirements.txt
click==8.0.3
Flask
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
Flask-WTF
flask_behind_proxy
flask_sqlalchemy
pytest==6.2.4
pycodestyle==2.7.0
flask-login
flask-bcrypt
requests==2.26.0
pandas==1.3.3
python-dotenv==0.19.1
